refactor: rename Robot folder to robot
