+++ source/controllers/tests/testDissipativeImpedanceController.cpp
@@ -1,8 +1,8 @@
 #include "controllers/impedance/Dissipative.hpp"
 #include "state_representation/Space/Cartesian/CartesianTwist.hpp"
 #include "state_representation/Space/Cartesian/CartesianWrench.hpp"
-#include "state_representation/Robot/JointVelocities.hpp"
-#include "state_representation/Robot/JointTorques.hpp"
+#include "state_representation/robot/JointVelocities.hpp"
+#include "state_representation/robot/JointTorques.hpp"
 #include <numeric>
 #include <gtest/gtest.h>
