+++ source/dynamical_systems/include/dynamical_systems/Linear.hpp
@@ -9,8 +9,8 @@
 #include "dynamical_systems/DynamicalSystem.hpp"
 #include "dynamical_systems/Exceptions/IncompatibleSizeException.hpp"
 #include "state_representation/parameters/Parameter.hpp"
-#include "state_representation/Robot/JointPositions.hpp"
-#include "state_representation/Robot/JointState.hpp"
+#include "state_representation/robot/JointPositions.hpp"
+#include "state_representation/robot/JointState.hpp"
 #include "state_representation/Space/Cartesian/CartesianPose.hpp"
 #include "state_representation/Space/Cartesian/CartesianState.hpp"
 #include "state_representation/Space/Cartesian/CartesianTwist.hpp"
