+++ source/dynamical_systems/src/DynamicalSystem.cpp
@@ -1,5 +1,5 @@
 #include "dynamical_systems/DynamicalSystem.hpp"
-#include "state_representation/Robot/JointState.hpp"
+#include "state_representation/robot/JointState.hpp"
 #include "state_representation/Space/Cartesian/CartesianPose.hpp"
 #include "state_representation/Space/Cartesian/CartesianState.hpp"
 #include "state_representation/Space/Cartesian/CartesianTwist.hpp"
