+++ source/state_representation/include/state_representation/robot/Jacobian.hpp
@@ -6,8 +6,8 @@
 #pragma once
 
 #include "state_representation/exceptions/IncompatibleSizeException.hpp"
-#include "state_representation/Robot/JointTorques.hpp"
-#include "state_representation/Robot/JointVelocities.hpp"
+#include "state_representation/robot/JointTorques.hpp"
+#include "state_representation/robot/JointVelocities.hpp"
 #include "state_representation/Space/Cartesian/CartesianTwist.hpp"
 #include "state_representation/Space/Cartesian/CartesianWrench.hpp"
 #include "state_representation/State.hpp"
