+++ source/state_representation/include/state_representation/robot/JointTorques.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "state_representation/Robot/JointState.hpp"
+#include "state_representation/robot/JointState.hpp"
 
 namespace state_representation {
 /**
