+++ source/state_representation/src/robot/JointPositions.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/Robot/JointPositions.hpp"
+#include "state_representation/robot/JointPositions.hpp"
 #include "state_representation/exceptions/EmptyStateException.hpp"
 #include "state_representation/exceptions/IncompatibleStatesException.hpp"
