+++ source/state_representation/src/robot/JointTorques.cpp
@@ -1,4 +1,4 @@
-#include "state_representation/Robot/JointTorques.hpp"
+#include "state_representation/robot/JointTorques.hpp"
 #include "state_representation/exceptions/EmptyStateException.hpp"
 #include "state_representation/exceptions/IncompatibleStatesException.hpp"
